requirements.txt
openai>=1.0
langchain>=0.1
python-dotenv>=1.0
